chore(localized-text-field): move to packages
