+++ packages/components/fields/localized-text-field/src/index.js
@@ -1,1 +1,3 @@
 export { default } from './localized-text-field';
+
+export { default as version } from './version';
