+++ packages/components/fields/localized-text-field/src/localized-text-field.visualroute.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { LocalizedTextField } from '@commercetools-frontend/ui-kit';
-import { Suite, Spec } from '../../../../test/percy';
+import { Suite, Spec } from '../../../../../test/percy';
 
 const value = {
   en: 'hello world',
